+++ ReactionTimeApp/View/TestViewComponents/TestStateViews/FalseStartView.swift
@@ -1,0 +1,18 @@
+//
+//  FalseStartView.swift
+//  ReactionTimeApp
+//
+//  Created by Toby Weir on 14/07/2025.
+//
+
+import SwiftUI
+
+struct FalseStartView: View {
+    var body: some View {
+        Text("you pressed too early")
+    }
+}
+
+#Preview {
+    FalseStartView()
+}
