+++ ReactionTimeApp/View/TestViewComponents/TestStateViews/WaitingForUserView.swift
@@ -1,0 +1,18 @@
+//
+//  WaitingForUserView.swift
+//  ReactionTimeApp
+//
+//  Created by Toby Weir on 14/07/2025.
+//
+
+import SwiftUI
+
+struct WaitingForUserView: View {
+    var body: some View {
+        Text("Press now!")
+    }
+}
+
+#Preview {
+    WaitingForUserView()
+}
