+++ ReactionTimeApp/View/TestViewComponents/TestStateViews/WaitingRandomTimeView.swift
@@ -1,0 +1,18 @@
+//
+//  WaitingRandomTimeView.swift
+//  ReactionTimeApp
+//
+//  Created by Toby Weir on 14/07/2025.
+//
+
+import SwiftUI
+
+struct WaitingRandomTimeView: View {
+    var body: some View {
+        Text("Wait")
+    }
+}
+
+#Preview {
+    WaitingRandomTimeView()
+}
